--- src/main/java/ec/com/technoloqie/api/customer/TecCustomerServiceApplication.java
@@ -1,13 +1,0 @@
-package ec.com.technoloqie.api.customer;
-
-import org.springframework.boot.SpringApplication;
-import org.springframework.boot.autoconfigure.SpringBootApplication;
-
-@SpringBootApplication
-public class TecCustomerServiceApplication {
-
-	public static void main(String[] args) {
-		SpringApplication.run(TecCustomerServiceApplication.class, args);
-	}
-
-}
