--- src/main/java/ec/com/technoloqie/api/customer/dto/CustomerDto.java
@@ -1,23 +1,0 @@
-package ec.com.technoloqie.api.customer.dto;
-
-import java.util.Date;
-
-import lombok.AllArgsConstructor;
-import lombok.Data;
-import lombok.NoArgsConstructor;
-
-@Data
-@AllArgsConstructor
-@NoArgsConstructor
-public class CustomerDto {
-	
-	private Integer id;
-	private String pass;
-	private String createdBy;
-	private Date createdDate;
-	private String modifiedBy;
-	private Date modifiedDate;
-	private Boolean status;
-	private PersonDto person;
-
-}
