--- src/main/java/ec/com/technoloqie/api/customer/dto/LoginDto.java
@@ -1,10 +1,0 @@
-package ec.com.technoloqie.api.customer.dto;
-
-import lombok.Data;
-
-@Data
-public class LoginDto {
-
-	private String usernameOrEmail;
-	private String password;
-}
