--- src/main/java/ec/com/technoloqie/api/customer/entity/UserType.java
@@ -1,74 +1,0 @@
-package ec.com.technoloqie.api.customer.entity;
-
-import java.io.Serializable;
-import java.util.Date;
-
-import jakarta.persistence.Column;
-import jakarta.persistence.Entity;
-import jakarta.persistence.GeneratedValue;
-import jakarta.persistence.GenerationType;
-import jakarta.persistence.Id;
-import jakarta.persistence.PrePersist;
-import jakarta.persistence.Table;
-import jakarta.persistence.Temporal;
-import jakarta.persistence.TemporalType;
-import jakarta.validation.constraints.NotEmpty;
-import lombok.AllArgsConstructor;
-import lombok.Getter;
-import lombok.NoArgsConstructor;
-import lombok.Setter;
-
-/**
- * Entidad que representa un tipo de usuario para la persistencia
- *
- */
-@Getter
-@Setter
-@NoArgsConstructor
-@AllArgsConstructor
-@Entity
-@Table(name="SFSTUSERTYPE")
-public class UserType implements Serializable{
-	
-	private static final long serialVersionUID = 1L;
-    
-	@Id
-	@GeneratedValue(strategy = GenerationType.IDENTITY)
-	@Column(name = "USERTYPEID", nullable=false, unique=true)
-    private Integer id;
-    
-    @Column(name="NAME", nullable=false, unique=true)
-    private String name; 
-    
-    @Column(name="DESCRIPTION", nullable=false)
-    private String description;
-    
-    @NotEmpty(message ="no puede estar vacio")
-	@Column(name="CREATEDBY",nullable=false)
-	private String createdBy;
-	
-	@Column(name="CREATEDDATE",nullable=false)
-	@Temporal(TemporalType.DATE)
-	private Date createdDate;
-	
-	@Column(name="MODIFIEDBY")
-	private String modifiedBy;
-	
-	@Column(name="MODIFIEDDATE")
-	@Temporal(TemporalType.DATE)
-	private Date modifiedDate;
-	
-	@Column(name="STATUS")
-	private Boolean status;
-	
-	// @OneToMany(fetch = FetchType.LAZY, cascade = CascadeType.ALL)
-	// @JoinColumn(name = "USERTYPEID")
-	// private Collection<User> users;
-
-	@PrePersist 
-	public void prePersist() {
-		createdDate = new Date();
-		status = Boolean.TRUE;
-	}
-	
-}
