--- src/main/java/ec/com/technoloqie/api/customer/repository/IUserTypeRepository.java
@@ -1,12 +1,0 @@
-package ec.com.technoloqie.api.customer.repository;
-
-import java.util.Optional;
-
-import org.springframework.data.jpa.repository.JpaRepository;
-
-import ec.com.technoloqie.api.customer.entity.UserType;
-
-public interface IUserTypeRepository extends JpaRepository<UserType, Integer> {
-	
-	Optional<UserType> findByName(String name);
-}
