+++ src/main/java/ec/com/technoloqie/fwk/security/api/TecFwkSecurityServiceApplication.java
@@ -1,0 +1,13 @@
+package ec.com.technoloqie.fwk.security.api;
+
+import org.springframework.boot.SpringApplication;
+import org.springframework.boot.autoconfigure.SpringBootApplication;
+
+@SpringBootApplication
+public class TecFwkSecurityServiceApplication {
+
+	public static void main(String[] args) {
+		SpringApplication.run(TecFwkSecurityServiceApplication.class, args);
+	}
+
+}
