+++ src/main/java/ec/com/technoloqie/fwk/security/api/dto/LoginDto.java
@@ -1,0 +1,10 @@
+package ec.com.technoloqie.fwk.security.api.dto;
+
+import lombok.Data;
+
+@Data
+public class LoginDto {
+
+	private String usernameOrEmail;
+	private String password;
+}
