+++ src/main/java/ec/com/technoloqie/fwk/security/api/entity/Customer.java
@@ -1,0 +1,74 @@
+package ec.com.technoloqie.fwk.security.api.entity;
+
+import java.io.Serializable;
+import java.util.Date;
+
+import jakarta.persistence.CascadeType;
+import jakarta.persistence.Column;
+import jakarta.persistence.Entity;
+import jakarta.persistence.GeneratedValue;
+import jakarta.persistence.GenerationType;
+import jakarta.persistence.Id;
+import jakarta.persistence.JoinColumn;
+import jakarta.persistence.OneToOne;
+import jakarta.persistence.PrePersist;
+import jakarta.persistence.Table;
+import jakarta.persistence.Temporal;
+import jakarta.persistence.TemporalType;
+import jakarta.validation.constraints.NotEmpty;
+import jakarta.validation.constraints.Size;
+import lombok.AllArgsConstructor;
+import lombok.Getter;
+import lombok.NoArgsConstructor;
+import lombok.Setter;
+
+@Getter
+@Setter
+@NoArgsConstructor
+@AllArgsConstructor
+@Entity
+@Table(name="CUSTOMER")
+public class Customer implements Serializable{
+	
+	private static final long serialVersionUID = 4711757581207299131L;
+
+	@Id
+	@GeneratedValue(strategy = GenerationType.IDENTITY)
+	@Column(name = "CUSTOMERID",nullable=false, unique=true)
+	private Integer id;
+	
+	@NotEmpty(message ="no puede estar vacio")
+	@Size(min=4, message = "debe tener una longitud minima de 5")
+	@Column(name="PASS",nullable=false)
+	private String pass;
+	
+	@NotEmpty(message ="no puede estar vacio")
+	@Column(name="CREATEDBY",nullable=false)
+	private String createdBy;
+	
+	@Column(name="CREATEDDATE",nullable=false)
+	@Temporal(TemporalType.DATE)
+	private Date createdDate;
+	
+	@Column(name="MODIFIEDBY")
+	private String modifiedBy;
+	
+	@Column(name="MODIFIEDDATE")
+	@Temporal(TemporalType.DATE)
+	private Date modifiedDate;
+	
+	@Column(name="STATUS")
+	private Boolean status;
+	
+	
+	@PrePersist 
+	public void prePersist() {
+		createdDate = new Date();
+		status = Boolean.TRUE;
+	}
+	
+	//@OneToMany(fetch = FetchType.LAZY, cascade = CascadeType.ALL)
+	//@JoinColumn(name = "CUSTOMERID")
+	//private Collection<Account> accountCol;
+	
+}
