+++ src/main/java/ec/com/technoloqie/fwk/security/api/repository/IPersonRepository.java
@@ -1,0 +1,9 @@
+package ec.com.technoloqie.fwk.security.api.repository;
+
+import org.springframework.data.jpa.repository.JpaRepository;
+
+import ec.com.technoloqie.fwk.security.api.entity.Person;
+
+public interface IPersonRepository extends JpaRepository<Person, Integer>{
+
+}
